Fix signed/unsigned comparison warning as error from r9421.
git-svn-id: http://skia.googlecode.com/svn/trunk@9422 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ samplecode/SampleApp.cpp
@@ -496,7 +496,7 @@ public:
         if (SkOSMenu::kMixedState != fSubpixelState) {
             paint->setSubpixelText(SkOSMenu::kOnState == fSubpixelState);
         }
-        if (0 != fHintingState && fHintingState < SK_ARRAY_COUNT(gHintingStates)) {
+        if (0 != fHintingState && fHintingState < (int)SK_ARRAY_COUNT(gHintingStates)) {
             paint->setHinting(gHintingStates[fHintingState].hinting);
         }
         return true;
